Refactored Layout folder to be simple
