+++ src/components/Layout/index.tsx
@@ -2,7 +2,7 @@ import React, { Suspense, useContext } from 'react'
 import { Box, Stack } from '@chakra-ui/react'
 import dynamic from 'next/dynamic'
 import { WagmiStatusContext } from '@/components/Wagmi/WagmiStatusContext'
-import Navbar from '../Navbar/Navbar'
+import Navbar from './Navbar/Navbar'
 
 const Footer = dynamic(() => import('@/components/Layout/Footer/Footer'), {
   suspense: true,
@@ -11,7 +11,7 @@ const DeferredNetworkErrorNotification = dynamic(
   () => import('@/components/Layout/Network/DeferredNetworkErrorNotification'),
 )
 const GoogleAnalyticsScripts = dynamic(
-  () => import('@/components/Layout/Layout/GoogleAnalyticsScripts'),
+  () => import('@/components/Layout/GoogleAnalyticsScripts'),
 )
 
 const Layout = ({
