+++ src/pages/_app.tsx
@@ -5,7 +5,7 @@ import '@/editor-plugins/pluginStyles.css'
 import { ChakraProvider, createStandaloneToast } from '@chakra-ui/react'
 import type { AppProps } from 'next/app'
 import { Provider as ReduxProviderClass } from 'react-redux'
-import Layout from '@/components/Layout/Layout/Layout'
+import Layout from '@/components/Layout'
 import SEOHeader from '@/components/SEO/Default'
 import { store } from '@/store/store'
 import Fonts from '@/theme/Fonts'
